gradio_code_search/requirements.txt
gradio
python-dotenv
loguru         # Or remove if using standard logging
httpx
qdrant-client>=1.7.0 # Use a recent version
GitPython      # For cloning GitHub repos
psutil         # For memory logging (optional but recommended)
tenacity       # For retries in embedding client